Fix issue preventing typing in the input field from the "Follow-up questions" area
The problem was the `event.currentTarget` used inside the closure of `setState((prev) => {...})`.

When used inside the closure, the value of `event.currentTarget` became `null`, as the target was not the current anymore (in this carried over event). But the `event.target` remained set. So saving the `event.currentTarget` in an constant before passing it to the closure solves the issue. But as the `event.target` was always set with the expected value, it was safer to use it instead.